requirements.txt
azure-identity
dateutils
flake8
msal
msgraph-sdk
pathvalidate
python-dateutil
requests
tqdm